requirements.txt
mkdocs>=1.4
mkdocs-material>=9.0
pymdown-extensions>=9.0
streamlit>=1.30
pandas>=2.0
numpy>=1.23
pyyaml>=6.0
pytest>=8.0

# Optional additional plugins you may add later:
# - mkdocs-minify-plugin
# - mkdocs-git-revision-date-localized-plugin
